Still does not scale
The part about finding a relative interior point is horrible.
Should be able to do with one linear program instead of roughly
d / 2.  Use the add epsilon as a variable method, similar to
the method used to calculate generic directions of recession
in section 3.13 of Geyer (2009, Electronic Journal of Statistics,
3, 259–289).